Add files via upload
The completed version of my Express-Jobly assignment. I have done a lot of troubleshooting the error message "TypeError: pathRegexp is not a function" when I run 'jest -i' with my mentor, Brandon, from Springboard Academy (including trying different versions of packages) but we were unable to find the cause of the error. After downloading the solution code and installing all the dependencies used in the solution then running 'jest -i', I got the same error message.